% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logcumulant-package.R
\docType{package}
\name{logcumulant-package}
\alias{logcumulant}
\alias{logcumulant-package}
\title{logcumulant: Goodness-of-Fit Tests and Diagrams Based on Mellin Log-Cumulants}
\description{
The \pkg{logcumulant} package implements a family of three complementary
goodness-of-fit tests based on an adaptation of Hotelling's \eqn{T^2}
statistic applied to vectors of sample log-cumulants (Mellin statistics) for
positive-support reliability data, together with three diagnostic diagrams.
}
\section{Main functions}{

\describe{
  \item{Diagrams}{\code{\link{log_cumulant_diagram}},
    \code{\link{kurtosis_diagram}}, \code{\link{cv_diagram}},
    \code{\link{three_diagrams}}, \code{\link{multi_lc_diagram}},
    \code{\link{plot_lc}}}
  \item{Tests}{\code{\link{T2_all}}, \code{\link{T2_bootstrap}},
    \code{\link{gof_compare_all}}}
  \item{Simulation}{\code{\link{size_study}}, \code{\link{power_study}}}
  \item{Building blocks}{\code{\link{mle_fit}}, \code{\link{theoretical_lc}},
    \code{\link{jacobian_J}}, \code{\link{fisher_closed}}}
}
}

\section{Data}{

Nine reliability datasets are bundled as \code{\link{reliability_datasets}}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/raydonal/logcumulant}
  \item \url{https://raydonal.github.io/logcumulant/}
  \item Report bugs at \url{https://github.com/raydonal/logcumulant/issues}
}

}
\author{
\strong{Maintainer}: Raydonal Ospina \email{raydonal@de.ufpe.br} (\href{https://orcid.org/0000-0002-9884-9090}{ORCID})

Authors:
\itemize{
  \item Raydonal Ospina \email{raydonal@de.ufpe.br} (\href{https://orcid.org/0000-0002-9884-9090}{ORCID})
}

}
\keyword{internal}
