% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reliability_datasets}
\alias{reliability_datasets}
\title{Nine reliability datasets used in the applications}
\format{
A named list of numeric vectors.
}
\usage{
data(reliability_datasets)
}
\description{
A named list with nine positive-valued reliability datasets analyzed in the
paper: Kevlar, Resistors, Tensile, Airplane, BallBearing, Airborne, Failure,
Yarn, AirCon.
}
\keyword{datasets}
